requirements.txt
anthropic>=0.39.0
zhipuai>=2.0.0
python-pptx>=1.0.0
edge-tts>=6.1.0
Pillow>=10.0.0
python-magic>=0.4.27
python-dotenv>=1.0.0
